requirements.txt

pyperclip
aiofiles
rich
wgpu
glfw
numpy
PyYAML
colored-traceback
pydantic